Run checks & tests in GitHub Actions on every push
Add Tests badge to README

While templating via ytt or a similar utility would be expected in this
situation, I want to challenge the assumption that using a utility to
template YAML is a better approach than copy-pasting & renaming the same
thing n times. We know how complicated templating turned out to be in
rabbitmq-ci after a few years of usage, let's try something else here.